+++ .gitignore
@@ -1,4 +1,3 @@
 weekly-report.md
 __pycache__
-cache/*
-!cache/.gitkeep
+cache.sqlite
