+++ requirements.in
@@ -1,2 +1,4 @@
 pip-tools
 argparse
+requests
+requests_cache
