+++ requirements.txt
@@ -2,9 +2,15 @@
 # This file is autogenerated by pip-compile
 # To update, run:
 #
-#    pip-compile requirements.in
+#    pip-compile
 #
 argparse==1.4.0
+certifi==2019.9.11        # via requests
+chardet==3.0.4            # via requests
 click==7.0                # via pip-tools
+idna==2.8                 # via requests
 pip-tools==4.1.0
+requests-cache==0.5.2
+requests==2.22.0
 six==1.12.0               # via pip-tools
+urllib3==1.25.6           # via requests
